refactor: move OAS compare functions and utils to compare/oas/; setupRouter to compare/oas/setup.ts
