+++ .gitignore
@@ -2,3 +2,5 @@ coverage/*
 dist/*
 node_modules/*
 .idea
+.vscode/
+src/compare/node_modules/
