+++ src/compare/index.ts
@@ -3,13 +3,13 @@ import type Router from "find-my-way";
 import type { HTTPMethod } from "find-my-way";
 import type { OpenAPIV2, OpenAPIV3 } from "openapi-types";
 import { compareAsyncInteraction } from "#compare/asyncapi/index";
-import { compareReqBody } from "#compare/requestBody";
-import { compareReqHeader } from "#compare/requestHeader";
-import { compareReqPath } from "#compare/requestPath";
-import { compareReqQuery } from "#compare/requestQuery";
-import { compareReqSecurity } from "#compare/requestSecurity";
-import { compareResBody } from "#compare/responseBody";
-import { compareResHeader } from "#compare/responseHeader";
+import { compareReqBody } from "#compare/oas/requestBody";
+import { compareReqHeader } from "#compare/oas/requestHeader";
+import { compareReqPath } from "#compare/oas/requestPath";
+import { compareReqQuery } from "#compare/oas/requestQuery";
+import { compareReqSecurity } from "#compare/oas/requestSecurity";
+import { compareResBody } from "#compare/oas/responseBody";
+import { compareResHeader } from "#compare/oas/responseHeader";
 import type { AsyncAPIDocument, Message } from "#documents/asyncapi";
 import { parse as parseAsyncapi } from "#documents/asyncapi";
 import { parse as parseOas } from "#documents/oas";
@@ -19,7 +19,8 @@ import type { Result } from "#results/index";
 import { baseMockDetails } from "#results/index";
 import { type Config, type ConfigKeys, DEFAULT_CONFIG } from "#utils/config";
 import { ARRAY_SEPARATOR } from "#utils/queryParams";
-import { setupAjv, setupRouter } from "./setup";
+import { setupRouter } from "./oas/setup";
+import { setupAjv } from "./setup";
 
 export interface ComparatorOptions {
   oas?: OpenAPIV2.Document | OpenAPIV3.Document;
